Sanitize reporter options from cli
